BG2-3016: Fix metacampaign ticker showing wrong number by using bigge…

+++ schema/Donation.sql
@@ -59,7 +59,7 @@ CREATE TABLE `Donation` (
   `payoutSuccessful` tinyint(1) DEFAULT NULL,
   `paymentCard_brand` varchar(255) DEFAULT NULL,
   `paymentCard_country` varchar(255) DEFAULT NULL,
-  `expectedMatchAmount_amountInPence` int NOT NULL,
+  `expectedMatchAmount_amountInPence` bigint NOT NULL,
   `expectedMatchAmount_currency` varchar(3) NOT NULL,
   PRIMARY KEY (`id`),
   UNIQUE KEY `UNIQ_C893E3F6D17F50A6` (`uuid`),
